Fix readonly bugs
For variables that capture the output of a shell
command, declaring them readonly on the same line
will prevent the status code from getting
captured correctly. Once it is set as readonly,
then, you cannot write to it. So, we first set
the variable and THEN set it as readonly. Problem solved.